Fix issues with gocean extraction driver creator
