+++ src/psyclone/psyir/nodes/extract_node.py
@@ -226,8 +226,7 @@ class ExtractNode(PSyDataNode):
 
         if self._driver_creator:
             nodes = self.children
-            region_name_tuple = self.get_unique_region_name(nodes, {})
-            region_name_tuple = (region_name_tuple[0], self._region_name)
+            region_name_tuple = self.get_unique_region_name(nodes)
 
             self.bring_external_symbols(read_write_info,
                                         self.ancestor(Routine).symbol_table)
@@ -301,7 +300,7 @@ class ExtractNode(PSyDataNode):
                 suffix += 1
         return postfix+str(suffix)
 
-    def get_unique_region_name(self, nodes, options):
+    def get_unique_region_name(self, nodes):
         '''This function returns the region and module name. If they are
         specified in the user options, these names will just be returned (it
         is then up to the user to guarantee uniqueness). Otherwise a name
@@ -310,52 +309,12 @@ class ExtractNode(PSyDataNode):
 
         :param nodes: a list of nodes.
         :type nodes: list of :py:obj:`psyclone.psyir.nodes.Node`
-        :param options: a dictionary with options for transformations.
-        :type options: Dict[str, Any]
-        :param (str,str) options["region_name"]: an optional name to \
-            use for this PSyData area, provided as a 2-tuple containing a \
-            location name followed by a local name. The pair of strings \
-            should uniquely identify a region unless aggregate information \
-            is required (and is supported by the runtime library).
 
         '''
-        # We don't use a static method here since it might be useful to
-        # overwrite this functions in derived classes
-        name = options.get("region_name", None)
-        if name:
-            # pylint: disable=too-many-boolean-expressions
-            if not isinstance(name, tuple) or not len(name) == 2 or \
-               not name[0] or not isinstance(name[0], str) or \
-               not name[1] or not isinstance(name[1], str):
-                raise ValueError(
-                    "Error in PSyDataTrans. The name must be a "
-                    "tuple containing two non-empty strings.")
-            # pylint: enable=too-many-boolean-expressions
-            # Valid PSyData names have been provided by the user.
-            return name
-
-        from psyclone.psyGen import InvokeSchedule, Kern
+        from psyclone.psyGen import InvokeSchedule
         invoke = nodes[0].ancestor(InvokeSchedule).invoke
         module_name = invoke.invokes.psy.name
-
-        # Use the invoke name as a starting point.
-        region_name = invoke.name
-        kerns = []
-        for node in nodes:
-            kerns.extend(node.walk(Kern))
-
-        if len(kerns) == 1:
-            # This PSyData region only has one kernel within it,
-            # so append the kernel name.
-            region_name += f"-{kerns[0].name}"
-
-        # Add a region index to ensure uniqueness when there are
-        # multiple regions in an invoke.
-        key = module_name + "|" + region_name
-        idx = self._used_kernel_names.get(key, 0)
-        self._used_kernel_names[key] = idx + 1
-        region_name += f"-r{idx}"
-        return (module_name, region_name)
+        return (module_name, self._region_name)
 
     # -------------------------------------------------------------------------
     @staticmethod
