+++ src/psyclone/tests/domain/common/extract_driver_creator_test.py
@@ -50,7 +50,8 @@ from psyclone.domain.gocean.transformations import (GOceanExtractTrans,
                                                     GOConstLoopBoundsTrans)
 from psyclone.psyir.nodes import Routine, Loop
 from psyclone.psyir.symbols import ContainerSymbol, SymbolTable
-from psyclone.psyir.transformations import PSyDataTrans
+from psyclone.psyir.tools import ReadWriteInfo
+from psyclone.psyir.transformations import PSyDataTrans, TransformationError
 from psyclone.tests.utilities import get_invoke
 
 # API names
@@ -265,7 +266,6 @@ def test_rename_suffix_if_name_clash():
     filename.
 
     '''
-    return  # TODO: This 'driver_test.f90' do not produce name clashes, the
     # test has to be completely redone
     # Use tmpdir so that the driver is created in tmp
     etrans = GOceanExtractTrans()
@@ -277,48 +277,18 @@ def test_rename_suffix_if_name_clash():
                                         ("module_name", "local_name")})
     extract_code = str(psy.gen)
 
-    # This kernel calls compute_kernel(out_fld%data, in_out_fld%data,
-    #                                  out_fld_post%data, dx%data)
-    # with the access patterns:
-    #   out_fld%data:      write
-    #   in_out_fld%data:   read+write
-    #   out_fld_post%data: read
-    # Due to the name clash of "out_fld_data"+"_post" (first parameter output
-    # value) and "out_fld_data_post" (third parameters as input argument)
-    # the _post suffix is changed to "_post0". So the file will
-    # contain:
-    #   out_fld_data: the input value of the written field, since even if a
-    #                 parameter is output only, its input value is stored
-    #                 in case that the kernel only updates some elements
-    #   in_out_fld_data: the input value of the in_out_fld argument
-    #   out_fld_post_data: the input value of the read-only array
-    # Then for the written fields:
-    #   in_out_fld_data_post0: the value of in_out_fld after the kernel
-    #   out_fld_data_post0: the output value of output_fld after the kernel
-    # Test that these fields are indeed given to the extraction library.
-    # Note that the last two ProvideVariable calls will actually be after
-    # the kernel call.
+    # This kernel calls compute_kernel(out_fld%data, out_fld_data%data,
+    #                                  out_fld_data_post%data, dx%data)
+    # Make sure the extraction does not clash the names of
+    # flattening out_fld%data to out_fld_data or out_fld_data_post
     expected_lines = [
-      # Declaration, first the three arguments
-      'CALL extract_psy_data % PreDeclareVariable("out_fld_data", \
-out_fld_data)',
-      'CALL extract_psy_data % PreDeclareVariable("in_out_fld_data", '
-      'in_out_fld_data',
-      'CALL extract_psy_data % PreDeclareVariable("out_fld_post_data", '
-      'out_fld_post_data)',
-      # Declare the two variables to be written after the kernel, with
-      # the suffix `_post0` added.
-      'CALL extract_psy_data % PreDeclareVariable("in_out_fld_data_post0", '
-      'in_out_fld',
-      'CALL extract_psy_data % PreDeclareVariable("out_fld_post0", out_fld)',
-      # Provide the variables before the kernel call:
-      'CALL extract_psy_data % ProvideVariable("out_fld", out_fld)',
-      'CALL extract_psy_data % ProvideVariable("in_out_fld", in_out_fld)',
-      'CALL extract_psy_data % ProvideVariable("out_fld_post", out_fld_post)',
-      # Provide the variables after the kernel call:
-      'CALL extract_psy_data % ProvideVariable("in_out_fld_post0", '
-      'in_out_fld)',
-      'CALL extract_psy_data % ProvideVariable("out_fld_post0", out_fld)']
+      'CALL extract_psy_data % PreDeclareVariable("out_fld_data_1", '
+      'out_fld_data_1)',
+      'CALL extract_psy_data % PreDeclareVariable("out_fld_data_1_post", '
+      'out_fld_data_1)',
+      'call compute_kernel_code(i, j, out_fld_data_1, out_fld_data_data, '
+      'out_fld_data_post_data, dx_data, out_fld_data_post_grid_dx, '
+      'out_fld_data_post_grid_gphiu)']
     for line in expected_lines:
         assert line in extract_code, line + "\n -- not in --\n" + extract_code
 
@@ -332,56 +302,14 @@ out_fld_data)',
         driver_code = driver_file.read()
 
     expected = """
-  real*8, allocatable, dimension(:,:) :: out_fld
-  real*8, allocatable, dimension(:,:) :: in_out_fld
-  real*8, allocatable, dimension(:,:) :: out_fld_post
-  real*8, allocatable, dimension(:,:) :: in_out_fld_post0
-  real*8, allocatable, dimension(:,:) :: out_fld_post0
-  call extract_psy_data%ReadVariable('in_out_fld', in_out_fld)
-  call extract_psy_data%ReadVariable('in_out_fld_post0', in_out_fld_post0)
-  call extract_psy_data%ReadVariable('out_fld_post0', out_fld_post0)
-  call extract_psy_data%ReadVariable('out_fld', out_fld)
-  call extract_psy_data%ReadVariable('out_fld_post', out_fld_post)"""
+  real*8, allocatable, dimension(:,:) :: out_fld_data_1
+  real*8, allocatable, dimension(:,:) :: out_fld_data_1_post
+  call extract_psy_data%ReadVariable('out_fld_data_1', out_fld_data_1)
+  call extract_psy_data%ReadVariable('out_fld_data_1_post', \
+out_fld_data_1_post)"""
 
     for line in expected.split("\n"):
-        assert line in driver_code
-
-    # Now test that more than one variable clash is handled. The third
-    # invoke uses:
-    # "out_fld" as output field
-    # "out_fld_post" as input field (first clash -->
-    # suffix becomes "_post0")
-    # "out_fld_post0" as input+output field (next clash -->
-    # suffix = "_post1")
-    psy, invoke = get_invoke("driver_test.f90",
-                             GOCEAN_API, idx=2, dist_mem=False)
-    schedule = invoke.schedule
-    # We don't check the driver, we already tested that the
-    # driver picks up the adjusted suffix above
-    etrans.apply(schedule.children[0])
-    extract_code = str(psy.gen)
-
-    # Check that *out_fld* is declared correctly: it is only declared as
-    # output value, but it must also be added as input field, to make sure
-    # we have all values of a field, even if the kernel doesn't update them:
-    assert 'PreDeclareVariable("out_fld_post1", out_fld)' in extract_code
-    assert 'PreDeclareVariable("out_fld", out_fld)' in extract_code
-
-    # Check that *out_fld_post* (input/output) is declared correctly. It
-    # must be declared twice: once for the input value using the original
-    # variable name, and once as output using the "_post1" suffix.
-    assert ('PreDeclareVariable("out_fld_post", out_fld_post)'
-            in extract_code)
-    assert ('PreDeclareVariable("out_fld_post_post1", out_fld_post)'
-            in extract_code)
-
-    # Check that *out_fld_post0* is declared correctly: as input-only
-    # variable it must be declared once for using the original variable
-    # name.
-    assert ('PreDeclareVariable("out_fld_post0", out_fld_post0)'
-            in extract_code)
-    assert ('PreDeclareVariable("out_fld_post0_post1", out_fld_post0)'
-            not in extract_code)
+        assert line in driver_code, line + "\n -- not in --\n" + driver_code
 
 
 # -----------------------------------------------------------------------------
@@ -511,47 +439,3 @@ def test_driver_creation_import_modules(fortran_reader):
     mod_func = all_symbols["mod_func"]
     assert str(mod_func) == ("mod_func: RoutineSymbol<UnresolvedType, "
                              "pure=unknown, elemental=unknown>")
-
-
-# -----------------------------------------------------------------------------
-# @pytest.mark.usefixtures("change_into_tmpdir")
-# def test_driver_node_verification():
-#     '''Test that the create() method verifies the node list it receives
-#     and only accept the valid parameters.
-
-#     '''
-#     # Use tmpdir in case that the call below does not raise an
-#     # exception, which would result in the driver being created
-#     # in the current directory.
-
-#     api = "gocean"
-#     _, info = parse(os.path.join(get_base_path(api), "driver_test.f90"),
-#                     api=api)
-#     psy = PSyFactory(api, distributed_memory=False).create(info)
-#     invokes = psy.invokes.invoke_list
-
-#     edc = ExtractDriverCreator()
-
-#     # Provide the nodes in the wrong order.
-#     # Invoke #3 has all in all three kernels:
-#     schedule = invokes[3].schedule
-#     read_write_info = ReadWriteInfo()
-#     with pytest.raises(TransformationError) as err:
-#         edc.create(nodes=[schedule.children[1],
-#                           schedule.children[2],
-#                           schedule.children[0]],
-#                    read_write_info=read_write_info, prefix="extract",
-#                    postfix="post", region_name=("file", "region"))
-#     assert ("Children are not consecutive children of one parent"
-#             in str(err.value))
-#     assert ("has position 0, but previous child had position 2."
-#             in str(err.value))
-
-#     # Provide nodes from different invokes:
-#     with pytest.raises(TransformationError) as err:
-#         edc.create(nodes=[invokes[3].schedule.children[1],
-#                           invokes[2].schedule.children[0]],
-#                    read_write_info=read_write_info, prefix="extract",
-#                    postfix="post", region_name=("file", "region"))
-#     assert ("supplied nodes are not children of the same parent."
-#             in str(err.value))
